# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CortexOS X API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PROJECTS ─────────────────────────────────────────────────────────────────
...
  /research/query:
    post:
      operationId: researchQuery
      tags: [research]
      summary: Run a RAG-based research query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchQueryInput"
      responses:
        "200":
          description: Research result with sources and insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchResult"
...
  /research/runs/{id}:
    delete:
      operationId: deleteResearchRun
      tags: [research]
      summary: Delete a saved research run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        summary:
          type: ["string", "null"]
        userId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - tags
        - createdAt
        - updatedAt
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
        status:
          type: string
        dueDate:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        userId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - priority
        - status
        - createdAt
        - updatedAt
    TaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        dueDate:
          type: string
        projectId:
          type: integer
      required:
        - title
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        dueDate:
          type: string

    # ─── PROJECT SCHEMAS ───────────────────────────────────────────────────────
...
    ResearchQueryInput:
      type: object
      properties:
        query:
          type: string
          minLength: 1
        includeNotes:
          type: boolean
      required:
        - query
    ResearchResult:
      type: object
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/ResearchSource"
        insights:
          type: array
          items:
            type: string
        relatedNotes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
      required:
        - answer
        - sources
        - insights
        - relatedNotes
    ResearchSource:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        relevance:
          type: number
      required:
        - title
        - excerpt
        - relevance
...