# lib/api-zod/src/generated/types/researchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CortexOS X API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Note } from './note';
import type { ResearchSource } from './researchSource';

export interface ResearchResult {
  answer: string;
  sources: ResearchSource[];
  insights: string[];
  relatedNotes: Note[];
}